lots of changes for vertex_buffer_t

+++ include/shiny/voodoo/resources.hpp
@@ -25,8 +25,8 @@ namespace voodoo {
 		read_write,
 	};
 
-	auto create_buffer(ID3D11Buffer**, gpu_access_t, cpu_access_t, unsigned long data_size, void* data = nullptr) -> void;
-	auto map(ID3D11Resource*, D3D11_MAPPED_SUBRESOURCE*, D3D11_MAP, unsigned int subresource) -> void;
+	auto create_buffer(ID3D11Buffer**, gpu_access_t, cpu_access_t, uint32_t data_size, void* data = nullptr) -> void;
+	auto map(ID3D11Resource*, D3D11_MAPPED_SUBRESOURCE*, D3D11_MAP, uint32_t subresource) -> void;
 	auto unmap(ID3D11Resource*, uint32_t subresource) -> void;
 
 //======================================================================
